Notification inactive user. Create new job for Quartz

+++ src/main/java/by/topolev/network/quartz/ImageCleanUpJob.java
@@ -1,5 +1,8 @@
 package by.topolev.network.quartz;
 
+import java.util.Date;
+import java.util.List;
+
 import javax.annotation.Resource;
 
 import org.quartz.Job;
@@ -8,6 +11,8 @@ import org.quartz.JobExecutionException;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
+import by.topolev.network.dao.UrlImageDao;
+import by.topolev.network.domain.UrlImage;
 import by.topolev.network.service.ImageService;
 
 public class ImageCleanUpJob implements Job {
@@ -16,14 +21,17 @@ public class ImageCleanUpJob implements Job {
 
     @Resource
     private ImageService imageService;
+    @Resource
+    private UrlImageDao urlImageDao;
+    
+    private int maxTimeExcitingUnusingUrlImage = 60000;
 	
 	@Override
 	public void execute(JobExecutionContext jobContext) throws JobExecutionException {
-		System.out.println("START");
-        /*if(imageService != null) {
-       	List<UrlImage> list = urlImageDao.findAll();
+		LOGGER.debug("Start job ImageCleanUpJob");
+        if(imageService != null) {
+        	List<UrlImage> list = urlImageDao.findAll();
         	for (UrlImage urlImage : list){
-       
         		if ((urlImage.isMarkUsing() == false) &&
        				((new Date()).getTime()-urlImage.getDataCreate().getTime() > maxTimeExcitingUnusingUrlImage)){
         			imageService.deleteImage(urlImage.getUrlImage());
@@ -31,23 +39,7 @@ public class ImageCleanUpJob implements Job {
         	}
         } else {
             LOGGER.info("The imageService is not injected!");
-        }*/
+        }
 	}
 	
 }
-
-//JobDataMap map = jobContext.getMergedJobDataMap();
-//UrlImageDao urlImageDao = (UrlImageDao) map.get("urlImageDao");
-//ImageService imageService = (ImageService) map.get("imageService");
-//int maxTimeExcitingUnusingUrlImage = (int) map.get("maxTimeExcitingUnusingUrlImage");
-//System.out.println(maxTimeExcitingUnusingUrlImage);
-//if (urlImageDao !=null){
-//	List<UrlImage> list = urlImageDao.findAll();
-//	for (UrlImage urlImage : list){
-//
-//		if ((urlImage.isMarkUsing() == false) &&
-//				((new Date()).getTime()-urlImage.getDataCreate().getTime() > maxTimeExcitingUnusingUrlImage)){
-//			imageService.deleteImage(urlImage.getUrlImage());
-//		}
-//	}
-//}
